[BUGFIX] Improve allowed MIME type handling, add vcf handling
The mime-db collection requires .vcf files to have the
MIME type "text/x-vcard", but PHP's `finfo_file` or
`mime_content_type()` reports the MIME type as
"text/vcard", which is the actual RFC6350 recommendation.

With this patch, the MIME type "text/vcard" is added (patched) to
the list of mime-db JSON file containing the file-extension to
MIME-type assignments. This will allow .vcf files to be recognized
for both "text/vcard" (IANA/RFC) and "text/x-vcard" (Apache) variants.

An upstream issue [1] has been created so that ideally the hotfix
will not be required anymore at some point. Depending on that
outcome we might need an upgrade wizard that maps existing
"text/x-vcard" sys_file entries to "text/vcard".

Editor's note matching a sad trombone: Yes, this might happen for other
formats as they are revealed to us. Yes, there is no comprehensive
collection of every MIME type with a matching example files to
easily verify the outcome of PHP's MIME mapping to the db.json
structure.

Additionally, this patch adjusts the output message
for mismatching MIME types to contain the expected
type in the flash message output, so that integrators
can more easily adjust their configuration like this:

```
$GLOBALS['TYPO3_CONF_VARS']['SYS']['FileInfo']['mimeTypeCompatibility']
  [<reportedMIMEType>]
  [<fileExtension>] = <expectedMIMEType>;
```

[1] https://github.com/jshttp/mime-db/issues/388

Resolves: #106916
Related: #106735
Related: #106772
Releases: main, 13.4, 12.4
Change-Id: Ife52d90a065b86d818d6fe39bf9df6940dac3076
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89879
Tested-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <garvin@hick.ing>