[BUGFIX] Correctly handle TextTableElement delimiter fallback
When a table content element has no table_delimiter field,
the first save fails.

A TCA database row's field values are either
- array, when the field is visible in the form
- simple type, when the field is not in the form
  (with its default value)

When saving the element the first time, the table_delimiter default value
is a string "124", causing the `$row['table_delimiter'][0]` to do a
substring operation. This leads to failure.

Upon saving the next time, the default value is an integer and all is
fine.

This patch fixes the problem by checking if the value is really an array
before trying to access the first array element.

Resolves: #106917
Releases: main, 13.4, 12.4
Change-Id: Ieb93ad18097ee466f84a39745d1f07a0a37cfdd5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89781
Reviewed-by: Christian Weiske <typo3-2024@cweiske.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Weiske <typo3-2024@cweiske.de>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <garvin@hick.ing>