[DOCS] Adjust indentation and headlines in EXT:lowlevel
Non-generic headlines are crucial to both external
search engines and our on-site elastic search.

Subheadlines in chapter "Command line" are excluded,
as they will be overhauled and restructured in a
follow up.

Anchors and therefore permalinks were missing in places, and are now fixed. Indentations are adjusted, too.

Resolves: #106923
Releases: main, 13.4, 12.4
Change-Id: Idafa617cae64b4483a0b0dec503940902f0fef84
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89821
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>