[TASK] Add "runTests.sh -s checkRstRendering..." subcommands for docs
This patch adds two possibilities to render documentation and
check for errors with the ./Build/runTests.sh helper script,
paving the way to utilize it in CI (nightlies).

The command can be either run to test-render docs in all
subdirectories:

> ./Build/Scripts/runTests.sh -s checkRstRenderingAll

and also to check a specific core extension key:

> ./Build/Scripts/runTests.sh -s checkRstRenderingSingle lowlevel

Enabling in pipeline will be done in a future change after errors
and warnings has been resolved.

Resolves: #106925
Releases: main, 13.4, 12.4
Change-Id: Ic36a8dfd1cca7ee3758f2b164828e3b68576c609
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89790
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>