[BUGFIX] Account for fake UTC-0 datetime strings in readonly elements
In order to pass in correct unix timestamps, the DatetimeElement
needs to account for fake UTC-0 offsets in datetime strings that
have been prepared and normalized to fake UTC-0 in native datetime
format provider `FormDataProvider/DatabaseRowDateTimeFields`.

Note: For TYPO3 v14 this has been fixed in #105549 by removing the fake
UTC-0 handling in favor of unqualified localtime, which does not require
this normalization.

Resolves: #106937
Releases: 13.4
Related: #105549
Change-Id: I9f18eab1901fbbfd996d1f08fabe1bb5319b278f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89814
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benjamin Franzke <ben@bnf.dev>