[FEATURE] Add "discard" command to DataHandler
DataHandler now has a new "command" called "discard"
to remove elements in a workspace and discard its work
forever.

Previously this was handled via the "clearWSID" and
"flush" action parameter within the "version" command,
acting as "subcommands".

The previous action parameters are kept without deprecation
as it was implemented in Testing Framework and is
still around for a while. Although not documented,
they are considered internal, they are still available.

Resolves: #107533
Releases: main
Change-Id: I141cbfbcda1604ff2c75217a582af69b676f6725
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89823
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Oli Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oli Bartsch <bo@cedev.de>