[BUGFIX] Prevent type error in undefined felogin settings
In case TypoScript/flexform settings for felogin are not
loaded, the checks for their values need to take possibly
unset array keys into account.

For this, null-coalescing is used to fallback to "null",
which then resolves to the falsy-elvis operator "?:" to
apply the actual default.

Checking "forgotLinkHashValidTime" is now additionally
secured to only contain numbers, so that "0" will properly
fall back to "12 hours".

Resolves: #106942
Releases: main, 13.4
Change-Id: Ifdd85c908c555663a20787fcd034d62922b3d0e8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89857
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Garvin Hicking <garvin@hick.ing>