[BUGFIX] Add missing page type to settings
During the initial development of the settings API we had a set of
must-haves defined. This included the page type as we set the goal
to have a better user experience configuring the system for the
integrator. This was defined as an essential part of driving adoption.

Due to time constraints, and lack of development power this did not
make it into the initial release of the LTS version – as more important
issues needed to be handled before the release.

The lack of this type raised huge UX and DX issues, as for example
the pagetree is not visible anymore to hover over nodes since
the settings editor is running without the context of a pagetree
that existed in the constant editor. On the DX side, we can
only declare integers and still have no idea when fetching the
setting what the meaning for this value is. One of the essential
points the settings API was introduced to solve.

For that reason we are extending the setting types with an
additional page type, and also move it to the current stable LTS.
Even this is released after the LTS version we feel that the
benefits will help users better optimize and utilize the
settings for their projects, in addition to help moving forward
to v14 once it is released as developers need to have the option
to make their settings work on 2 major versions.

This type currently implements a page browser to allow the user to
select a page for a specific setting, while still displaying the UID
in the field. The user can now simply select a page instead of
having to remember the page UID as before.

This behavior will degrade gracefully when the type is used
outside of the backend scope and for example is for some
reason used in the "install tool" context. It will simply
not render the page browser as intended and fallback to its
underlying integer storage.

Usage Example:

  plugin.my_plugin.settings.storagePid:
    category: myplugin
    type: page
    default: 0

We are aware that this is still not the final form of this type.
There are still issues to be solved, like making the Reference-Index
aware of these references to, for example, block pages from
deletion when they are referenced in the settings, to avoid
misconfiguration.

With that in mind, this is something we are not yet sure how to
solve, and we felt it was too intrusive for now to pursue —
it needs more thought. For now, we are only implementing the meaning
for the DX and the convenience for the UX.

Resolves: #106953
Releases: main, 13.4
Change-Id: I763e6e66cff43c9d007f6ab2f11b1dfe5aeb615a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90013
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>