[TASK] Deprecate User TSConfig option auth.BE.redirectToURL
The User TSConfig option `auth.BE.redirectToURL` allows to
redirect a backend user after a successful backend login
to a configured url. This has been used in the past to
redirect a backend user to the frontend to use
frontend editing.

The User TSConfig option has been marked as deprecated with
this patch and will be removed in TYPO3 15.0.

Resolves: #106969
Releases: main
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: Ie723c1464d012aaa918e423f761daac8eb85bddf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89877
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>