[BUGFIX] Use table alias in WorkspaceRestriction
The WorkspaceRestriction misses a table alias in one
place which leads to "ambiguous" und thus failing
SQL when a table is joined that has column
"t3ver_state" as well.

The patch turns the existing unit tests into functional
tests and adds a case to verify the fix.

Resolves: #104548
Resolves: #102613
Releases: main, 13.4
Change-Id: I3daa198d96a4309d1c6b1c663a2e0e7717af1042
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89905
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>