[BUGFIX] Respect `#[IgnoreValidation]` attribute in controllers only
The `#[IgnoreValidation]` attribute can be used to disable a previously
configured validation for a property or method. In contrast to the
`#[Validate]` attribute, it was previously possible to attach an
`#[IgnoreValidation]` attribute to every class method. Since it is only
relevant in controller context and to streamline the behavior with the
`#[Validate]` attribute, `#[IgnoreValidation]` attributes are no longer
respected in non-controller class methods.

Resolves: #107235
Releases: main
Change-Id: I6e87d9d5db3e46758c5136f55eac3469e25b767d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89891
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>