[BUGFIX] Fix filetree filtering only expanding first level
Opening the `File > Filelist` with collapsed folders, and then
searching for a folder name only expanded the tree up to
the first level.

This is now fixed by expanding the tree up to the depth of the
returned tree children, by adding the 'loaded' metadata to it.

Resolves: #106683
Related: #93426
Releases: main, 13.4
Change-Id: Ic68b423cf501381ad2bd22301bb6a2694c6dce79
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89895
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Tested-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Garvin Hicking <garvin@hick.ing>