[TASK] Always reconstitute TSFE in PageContentErrorHandler
When calling persistAll EXT:solr will initialize all Sites.
Sites with root pages of type Shortcut to restricted pages
will trigger PageNotFound handling.

Before this change, given exceptionalErrors, if an error
was encountered it would throw an exception which was caught
too late and TSFE was lost.

With this change, regardless of exceptions, TSFE will be kept.

Resolves: #106984
Releases: main, 13.4
Change-Id: Icd8aef3d9c9c49414571e9eba998667161aa0049
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89914
Tested-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>