[BUGFIX] Correct usage of IsFileSelectableEvent
During the rework of the FileList, the IsFileSelectableEvent was not
migrated correctly. Instead of remaining available only in the
FileBrowser, it was moved to the FileList. Additionally, its purpose and
usage were misunderstood—likely due to the challenge of naming things
clearly.

This resulted in patches that attempted to fix the original issue by
applying workarounds, without addressing the root cause. Unfortunately,
these patches also introduced new regressions, further complicating the
behavior and stability of the FileList.

The underlying concept of the FileList assumes that each resource has a
primary action, which must be implemented by the consuming view.
Displaying disabled items was never intended.

The event was introduced to provide fine-grained control over which
files a user may select. This control should reflect the FileList's
display logic. However, it was mistakenly applied to the selectable
option, which has a different meaning: a file that is not selectable
should never be shown to the user in the first place.

This patch removes the related workarounds in the FileList and
introduces a new matcher, since the current matching logic only supports
OR operations. The event has now been relocated and is exclusively used
in the FileBrowser, as originally intended.

Related: #106560
Resolves: #106818
Releases: main, 13.4
Change-Id: I7c5ecd23d7dd71915bb5dcf7221c2d47f995fa7b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89624
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: core-ci <typo3@b13.com>