[FEATURE] Add event to change the record data in list view
A new event has been added to change the record data in the
list view.

Caution: Even though it is discouraged to hand in the
"parent object" (in this case DatabaseRecordList) the current
implementation encourages this as an exception.

Why? DatabaseRecordList is still full with public properties
that are needed from the outside, and also couples the
rendering, which should be changed at a later point,
making this Event "obsolete" as well.

Resolves: #107003
Resolves: #106206
Releases: main
Change-Id: Ib622acd6f414a8f05082e476a80943f8e4f80a80
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89941
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>