[BUGFIX] Optimize queries in RootlineUtility
When deploying a v13 project into real-life,
one quickly notices empty rootline caches and
the limitations of the database on
enterprise-level projects.

This is partially due to some limitations of
current Doctrine DBAL 4.2 (4.3 is much better),
but a runtime cache for the DB fields is much
better and faster due to unneeded quoting.

Resolves: #107149
Releases: main, 13.4
Change-Id: Id0d7af517f5183bd287fd04ee73a0f306634cc9c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90048
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>