[TASK] Add "RTE-Styleguide" preset showcasing specific RTE behaviour
This patch adds a "RTE-Styleguide" preset and uses it for

* `Form engine elements - rte` -> `rte_3`
* `Form engine elements - rte child inline_1` -> `rte_1`

When editing a "Form engine elements - rte" element in the tab
"in inline" the bug reported in #107093 can be reproduced,
the styles dropdown has a cut-off.

Having a distinct EXT:styleguide specific RTE preset can
help in the future to showcase things that otherwise should
not be implemented into the "Full" core preset.

Resolves: #107096
Related: #107093
Releases: main, 13.4
Change-Id: I637cea5ac5fe897249d6970eaa2e0841ce29dd85
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90183
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>