[TASK] Improve return type handling in HTTP/Stream
PHP's `stream_get_contents()` function has a return type decleration
of "string|false". The "false" return of that function can only
happen, if the `$offset` parameter of `stream_get_contents()` is set
to a value larger than `0`. The default is `-1`, so the TYPO3 code
could never trigger a 'false' condition and throw an exception.

With this patch, no exception is thrown (because it will never happen),
and the implicit function arguments are made explicit to be better
aware of this fact.

This will allow to upgrade to PHPStan 2.1.18 without touching the
baseline (only for the 12.4 baseline, as this patch is recommended
to not be backported to the security-fix-only branch).

Resolves: #107109
Related: #107110
Releases: main, 13.4
Change-Id: I8969ba0b6882ea68004c9165e18f9fd6cad5557e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90092
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>