[BUGFIX] Disable indexing of seo_sitemap
It currently is possible to call /sitemap.xml not only for the root page but for all pages, as it is just a page type. These sitemaps will be indexed and shown as search results by EXT:indexed_search. This change disables the indexing of those sitemaps.

Resolves: #107114
Related: #105837
Releases: main, 13.4, 12.4
Change-Id: I760d1f9959db44d4b3bc3a2a84876eb9c9ba6ebe
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90124
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>