[BUGFIX] Offload decorative thumbnail generation in file list
The file list currently generates a thumbnail for each "thumbnailable"
asset, used for drag&drop only in its current state. Many assets get
generated while they are not directly visible to the user, leading to
an additional overhead in IO, which drastically adds up on remote
storages.

To avoid premature generation of such assets, the generation is
offloaded to a PHP endpoint called on-demand when the new web
component `typo3-backend-thumbnail` is rendered in DOM.

Resolves: #106902
Releases: main, 13.4, 12.4
Change-Id: I49e3f7e73de0cfc0ba94c860a53e941f41634f4d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90118
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>