[BUGFIX] Utilize DuplicationBehavior ENUM in EXT:form
With the removal of nun-ENUM DuplicationBehavior values in
issue #105425 a leftover remained in the EXT:form
UploadedFileReferenceConverter. This was hard to spot
because it referenced only a "conflictMode" name instead
of "DuplicationBehavior".

With this patch the proper casting from a string to the
ENUM is performed, preventing a PHP TypeError when uploading
files through EXT:form:

```
TYPO3\CMS\Core\Resource\Folder::addUploadedFile(): Argument #2
($conflictMode) must be of type
TYPO3\CMS\Core\Resource\Enum\DuplicationBehavior, string given
```

Resolves: #107134
Related: #101151
Related: #105377
Related: #105425
Releases: main
Change-Id: Ib9513ce2f195384e1befc780a2a11a2f129f505a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90134
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Josua Vogel <j.vogel97@web.de>
Reviewed-by: Josua Vogel <j.vogel97@web.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>