[BUGFIX] Ensure non-hidden translated records are renderd in Extbase
When requesting a translated record (in a workspace)
that has a default language record which is hidden,
the default language is not included in the
"$defaultLanguageRecordsSubSelect" because it is
selected as hidden etc. and the DefaultRestrictionContainer
was used in general.

This is now removed as the selection for the IN() queries
do not use restrictions anymore.

Resolves: #92442
Releases: main, 13.4
Change-Id: I9647a728284418a1ef9921588275661aa63ecb37
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90143
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Lisa Kreitz <lisamariakreitz@web.de>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Lisa Kreitz <lisamariakreitz@web.de>