[BUGFIX] Allow form element identifier in email finisher overrides
When using the EXT:form plugin as a content element, it's flexform
definition allows to override certain elements of the form config.

For example, the "Email to sender" finisher allows to specify
an email address, which FormEngine renders as "type: email". Thus,
this only allows to enter valid emails.

However, one main reason to use flexform overrides for the plugin is
to be able to insert form element references so that the "Email
to sender" plugin actually uses the input of a frontend form
to send the mail to, by inserting `{userMailto}` or something else
as the identifier.

Doing so prevented FormEngine validation to save the input because
it no longer matched an email.

Sadly, the "type: email" attribute is tightly bound to the EXT:form
backend GUI. If just set to "type: input" then the input of the backend
would no longer enforce only entering valid emails. EXT:form extracts
the flexform settings from the base YAML with a complex method
to provide editing a plugin instance's specific mail addresses
within FormEngine.

So a solution was needed that allowed to:

1) Require to enter only valid emails in EXT:form GUI
2) Allow to enter either a valid email in FormEngine flexform
overrides, or a `{...}` identifier reference.

This is now solved with a custom formeval implementing a check
for both syntaxes, and not storing invalid values.

Resolves: #107025
Releases: main, 13.4
Change-Id: Ic3daeb941d5d2cc991e9f4828b6c2973ab3e0a8a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90729
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>