[TASK] Enhance Extbase file upload consistency checks
The file upload handling has a `MimeTypeValidator` which internally
also performs a file extension to mime-type consistency check.
However, in case the `mimeType` validator is not explicitly set,
the consistency check won't be executed either.

This change introduces the `FileExtensionMimeTypeConsistencyValidator`
which performs a generic file extension to mime-type consistency check
(in contrast, `MimeTypeValidator` would compare the assumed mime-types
to the configured allowed mime-types additionally).

In addition the `FileExtensionValidator` is added, allowing to check
for specific file extensions. This is considered a missing feature.

Resolves: #107191
Related: #107142
Releases: main, 13.4
Change-Id: I8351da0220d9c3768f445b3b763b25ec508d18e8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90219
Tested-by: Torben Hansen <derhansen@gmail.com>
Reviewed-by: Torben Hansen <derhansen@gmail.com>
Tested-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Garvin Hicking <garvin@hick.ing>