[BUGFIX] Properly parse Country property in CountrySelectViewHelper
The implementation of #99911 missed to address that the
`<f:form.countrySelect>` ViewHelper needs to properly evaluate
the Country object, when referenced as a domain model property
instead of getting accessed via `value`.

This patch adds a test for this case and adapts retrieval
in the ViewHelper's `convertToPlainValue()` method.

Also, a drive-by typo is fixed.

Resolves: #107200
Related: #99911
Related: #104997
Releases: main
Change-Id: Ibf202c21a70d96303b188d7a5ad4fd45e055f263
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90238
Tested-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Tested-by: Simon Praetorius <simon@praetorius.me>
Tested-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Garvin Hicking <garvin@hick.ing>