[TASK] Skip fixture extensions for root composer.json integrity check
Different tools and scripts has been put in place to keep the quality
of the project up and includes for example a integrity check for the
root composer.json that all dependencies are also required already
skipping system extensions.

While it is still a good thing to check composer.json files of any
test fixture extension it is not wanted to have them required for
TYPO3 Core Development and was not an issue until now. To introduce
tests using two test fixture extensions, where one of them requires
the other as a dependency, it would be reported that this require is
missing in the root `composer.json` considered as false-positve.

Glady, some time back the decision was made to use `typo3tests/` as
composer package name prefix for test fixture extensions making it
now possible to skip them for the root composer.json check like it
has been done for the system extensions already.

Resolves: #107265
Related: #107231
Releases: main, 13.4, 12.4
Change-Id: I974a6e4be22c983a48040d9ebcfac6c503f39cbf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90342
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>