[BUGFIX] Ensure invalid pages should not stop DataHandler
When working in Workspaces, with redirects or
other parts where an editor does not have access
to a page, BE_USER->isInWebMount throws a RuntimeException.

This is really bad when dealing with workspaces,
e.g. interrupting the publishing process.

The exception is caught in this one place, and continued
further.

Resolves: #107103
Releases: main, 14.3
Change-Id: Idd97aba049114c733f3b5f3367ba7845a6b081db
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90384
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>