[!!!][TASK] Remove TypoScript getTSFE() condition function
To proceed with the removal of the TypoScriptFrontendController
god class, a few integrator related features have to be removed
before 14.0.

This change is about the 'getTSFE()' TypoScript condition
function which can be substituted with different conditions
quite easily.

While it would be relatively easy to create a b/w compat layer
that does not depend on TypoScriptFrontendController for this
function to continue to work, the decision is to remove it
entirely: We want to fully get rid of the keyword "TSFE" to
avoid this unfortunate wording, so the removal involves the
TypoScript references as well.

Resolves: #107473
Releases: main
Change-Id: I194c0f00c3939f0cb5aed594bc128ac9ef06ec84
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90702
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Garvin Hicking <garvin@hick.ing>