[BUGFIX] Consider already existing _computed in raw record
This patch adds support to read and build ComputedProperties
back from the _computed array.

There are cases where the raw record passed to RecordFactory
comes from the RawRecord->toArray() result. This array does
not possess the old computed property keys, but the newer
_computed array.

An example for such a case is:
- User adds page-content data processor to fetch records
- User renders each Record in column with f:cObject
- The rendering definition uses RecordFactory again to resolve relations
  from the data array passed to the ViewHelper.

The second RecordFactory call would result in wrong relations being
resolved if _computed is not considered.

Resolves: #107195
Related: #106303
Releases: main, 13.4
Change-Id: I724322d9c9c413dcb3c66fda522c16619282826b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90846
Tested-by: Oli Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oli Bartsch <bo@cedev.de>