[TASK] Extend allowed list of uploadable `mediafile_ext` files
The default configuration for
`$GLOBALS['TYPO3_CONF_VARS']['SYS']['mediafile_ext']`
has been enhanced with a list of formats commonly uploaded into
web-based CMS systems like TYPO3:

*  3gp
*  aac
*  aif
*  avif
*  heic
*  ico
*  m4a
*  m4v
*  mov
*  psd

This list is only enhanced for systems where the said configuration
option has not been customized already.

Enhancing this list allows to upload files with these extensions into
the file list (and use these files in TCA fields with
"common-media-types"), when the security flag
`security.system.enforceAllowedFileExtensions` is enabled.

This can be a relevant change, if these file types are now used in
records or content elements, that might not expect files of these types
for further operation (like frontend display).

Integrators need to ensure that all uploaded file types are handled
accordingly (e.g. embedding "mov" as a video, "ico" as an image, and
"psd" as a download), or have an appropriate fallback.

The `<f:media>` ViewHelper for example would iterate possible
`Renderer` implementations, that can deal with several of the
mentioned files (their MIME types), and try to render a file as an
image as fallback.

Due to this, the patch is only aimed at main.

Resolves: #107399
Related: #107033
Related: #106772
Related: #106735
Related: #106916
Related: #107398
Releases: main
Change-Id: I8aaddd63325ac273cf4adc36590f1faa3f623f77
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90577
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>