[FEATURE] Support more hashing algos in FAL
Previously, FAL's `LocalDriver` only supported `md5` and `sha1` as
hashing algorithms. While this may be good enough, it might be necessary
to use different hashing algorithms, depending on the use-case.

The method `LocalDriver->hash()` is now able to use any hashing
algorithm that is registered in PHP itself by building a `HashContext`
object and updating it by streaming the file content.

Resolves: #107441
Releases: main
Change-Id: I88a8e8372db09aa4f26c7f964f3c495148479928
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90668
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>