[BUGFIX] Prevent race condition on element access in Backend Login
user-pass-login attaches some events and functionality to the inputs
and buttons on the login screen. If a system is particularly slow,
this fails, or is done while a user is typing something, causing
invalid element access, which causes the JS to halt and preventing
backend login entirely.

The fix is to simply wait for the document to be ready and only then
accessing the inputs and attaching the events to them.

Resolves: #107067
Releases: main, 13.4
Change-Id: Id94a1b1de7c985ed466e3a5202f13de4fc32d82a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90677
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Pfaff