[FEATURE] Introduce scheduler task wizard
Implements a new wizard for creating scheduler
tasks in the Scheduler module to replace the
previous dropdown-based task selection.

The wizard provides a modern, categorized
interface similar to the content element
wizard with enhanced user experience. As
this provides a search functionality to
filter available task as well as the
option for default values.

This also prevents broken records by pre-selecting
task types instead of changing them in FormEngine.

The new wizard can be modified by the
new PSR-14 `ModifyNewSchedulerTaskWizardItemsEvent`.

Resolves: #107471
Releases: main
Change-Id: If6ddeaed2ecd8b9884eec59a4bfaea58992c3ec6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90701
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Oli Bartsch <bo@cedev.de>