[BUGFIX] Assure working redirect on discarded extension upload
When an extension gets uploaded in extension manager, an initial
dependency check is performed. If the uploaded extension defines any
unresolved dependencies, the request is redirected to the dependency
overview in `ListController::unresolvedDependenciesAction`. Since this
action requires POST requests, the redirect must be performed as 307
instead of 303 in order to forward the original HTTP request method.

Resolves: #107480
Releases: main, 13.4, 12.4
Change-Id: I2fb65e895362dca7e62943a44998aff1f473bbb4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90710
Reviewed-by: Torben Hansen <derhansen@gmail.com>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Josua Vogel <j.vogel97@web.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Torben Hansen <derhansen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Josua Vogel <j.vogel97@web.de>
Reviewed-by: Benni Mack <benni@typo3.org>