[BUGFIX] Avoid double resolving of FlexForm data structure
Previously, `FlexFormAjaxController` manually
resolved a flex form data structure based on
a given identifier and passed the result to the
`FormDataCompiler`. However, since the compiler
calls the form data providers — in particular
`TcaColumnsOverrides` — the manually resolved
structure was overwritten again. This caused
a logical error in `TcaFlexPrepare`.

The issue is now resolved by removing the manual
resolving in `FlexFormAjaxController`, as this
step is anyways handled by `TcaFlexPrepare`.

Besides fixing the bug, this also simplifies
the code and improves the overall flow.

Resolves: #107493
Releases: main
Change-Id: Ib2f6867a268e6b49bdccec69e84a76ff755f7593
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90736
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Oli Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>