[TASK] Remove further TSFE methods
The entire class is marked @deprecated already, usages of
TypoScriptFrontendController and global TSFE are set up
to be detected by the extension scanner already. A final
ReST file will be added when the class is fully dissolved.

The patch removed a couple of methods that have no internal
usage:
- logDeprecatedTyposcript() could be copied in-place if needed
again
- set_no_cache() has a substitution via request attribute
- set_cache_timeout_default() unused and outdated

Resolves: #107506
Releases: main
Change-Id: Ie7711441bcb2cc4b3d0492fd292c57446989c779
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90740
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>