[BUGFIX] Allow separate scrolling in new record wizard
The "New record" wizard modal popup has separate navigation
and content panes which previously scrolled in conjunction with
each other.

This patch now allows distinct scrolling in both areas, so the
contents of the selected navigation pane item can remain visible.

This is done by using a reflected property in the new record wizard
web component on-build, which is sent to its parent, so it can be
determined by the parent if the child has a navigation element. If
so, the parent can be set to overflow: hidden to allow the
navigation and the content to get individual overflow values.

Since it is timed in the willUpdate-Hook, this will also work correctly, if any events trigger on runtime. This can be beneficial in future
updates for events like this.

The mobile handling (which shows a dropdown instead of a navigation
pane) remains unaffected.

Resolves: #107477
Releases: main, 13.4
Change-Id: I0ca5a748240bd3d2217cf15e841a13eedd8ab26f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90747
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Garvin Hicking <garvin@hick.ing>