[TASK] Add documentation RST parsing to nightly CI pipeline
The script:

  Build/Scripts/runTests.sh -s checkRstRenderingAll

is now part of the nightly CI process. We should watch
if/when/how often it might fail to decide whether to
include this in the pre-merge at some point. This would
however add a minute or so to each CI run, which might
not be needed at this point.

Resolves: #107514
Releases: main, 13.4, 12.4
Change-Id: I1a6ef8cec31eefc9eebb09bb71631227df608ff5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90764
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>