[TASK] Remove "en" fallback in LocalizationFactory
LocalizationFactory is now simplified.

It previously returned getParsedData() with the
language keys "en" and the given $languageKey
as an array of labels.

In addition, it used internal calls with runtime
caches, even though this is not needed anymore, mainly
because
1. XML overrides are now covered properly in system cache
2. consumer code ($GLOBALS[LANG]) fetched "en" separately
   already.
3. Runtime cache is already in the comsumer code.

This allows us to thin out LocalizationFactory,
removing the runtime cache in this class, making this
class effectively stateless.

Resolves: #107517
Related: #107436
Releases: main
Change-Id: I055b2eb7feeb1b7295759262d6b40d69214f6fb4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90755
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>